Core: Container::Pop now with pick allows Poping containers
git-svn-id: svn://ultimatepp.org/upp/trunk@10627 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Vcont.hpp
@@ -588,7 +588,7 @@ void Array<T>::Insert(int i, const Array& x, int offset, int count) {
 template <class T>
 void Array<T>::Insert(int i, std::initializer_list<T> init)
 {
-	vector.InsertN(i, init.size());
+	vector.InsertN(i, (int)init.size());
 	for(auto q : init)
 		vector[i++] = new T(q);
 }
